Consolidate user validations, rename serializer
